add ios splash screen color, start star filter button
